Search feature
Added a search feature that will allow the user to search for a vehicle with the registration plate, currently still working on improving the search feature, it only notifies the user whether the vehicle exists or not, does not show the specific vehicle and its details. Will be working to solve this issue.